created quote table

+++ springboot/src/main/java/ca/jrvs/apps/trading/model/domain/IexQuote.java
@@ -127,11 +127,11 @@ public class IexQuote {
   @JsonProperty("iexBidPrice")
   private Double iexBidPrice;
   @JsonProperty("iexBidSize")
-  private Long iexBidSize;
+  private Integer iexBidSize;
   @JsonProperty("iexAskPrice")
   private Double iexAskPrice;
   @JsonProperty("iexAskSize")
-  private Long iexAskSize;
+  private Integer iexAskSize;
   @JsonProperty("marketCap")
   private Long marketCap;
   @JsonProperty("week52High")
@@ -490,12 +490,12 @@ public class IexQuote {
   }
 
   @JsonProperty("iexBidSize")
-  public Long getIexBidSize() {
+  public Integer getIexBidSize() {
     return iexBidSize;
   }
 
   @JsonProperty("iexBidSize")
-  public void setIexBidSize(Long iexBidSize) {
+  public void setIexBidSize(Integer iexBidSize) {
     this.iexBidSize = iexBidSize;
   }
 
@@ -510,12 +510,12 @@ public class IexQuote {
   }
 
   @JsonProperty("iexAskSize")
-  public Long getIexAskSize() {
+  public Integer getIexAskSize() {
     return iexAskSize;
   }
 
   @JsonProperty("iexAskSize")
-  public void setIexAskSize(Long iexAskSize) {
+  public void setIexAskSize(Integer iexAskSize) {
     this.iexAskSize = iexAskSize;
   }
